Split individual input plugins into separate modules.
